Add Carbon library to composer.json and update templates/reclamationsh/index.html.twig
The diff adds the Carbon library to the composer.json file and makes updates to the templates/reclamationsh/index.html.twig file. The specific changes to the file are not visible in the diff, but it appears that a link to a favicon has been added, as well as some CSS and JS files. Additionally, some HTML has been added to the file, including a header, a section with an image and text, and a footer.